refactor: use campus.config instead of campus.client.config
